refactor: dedupe date-key formatting, day-start and pagination helpers
- Add toLocalDateKey(value?) to date-utils and route the 12 hand-rolled
  'YYYY-MM-DD' local-date formatters through it (analytics-stats,
  dose-schedule-service, use-push-schedule-sync, correlation-chart,
  time-range-selector, schedule-view, compound-card-expanded,
  prescription-card, use-titration-drawer-form). Behaviour-identical.
- Move the duplicated getDayStartTimestamp (byte-identical in intake-service
  and use-intake-queries) to its canonical home in date-utils; both now
  import it, and use-intake-queries re-exports it for existing consumers.
- Move the PaginatedResult<T> interface (declared identically in
  intake-service and health-service) into service-result; both import it.
- Scope the new intake-service rollover test's fake timers to Date only so
  fake-indexeddb transactions don't stall.
