+++ src/components/analytics/correlation-chart.tsx
@@ -12,6 +12,7 @@ import {
 } from "recharts";
 import type { CorrelationResult, DataPoint } from "@/lib/analytics-types";
 import { cn } from "@/lib/utils";
+import { toLocalDateKey } from "@/lib/date-utils";
 
 // ---------------------------------------------------------------------------
 // Props
@@ -58,8 +59,7 @@ function coefficientColor(
 }
 
 function dayKey(ts: number): string {
-  const d = new Date(ts);
-  return `${d.getFullYear()}-${String(d.getMonth() + 1).padStart(2, "0")}-${String(d.getDate()).padStart(2, "0")}`;
+  return toLocalDateKey(ts);
 }
 
 function shortDate(ts: number): string {
