+++ src/lib/health-service.ts
@@ -1,5 +1,5 @@
 import { db, type WeightRecord, type BloodPressureRecord } from "@/lib/db";
-import { ok, err, type ServiceResult } from "@/lib/service-result";
+import { ok, err, type ServiceResult, type PaginatedResult } from "@/lib/service-result";
 import { generateId, syncFields } from "@/lib/utils";
 import { writeWithSync } from "@/lib/sync-queue";
 import { schedulePush } from "@/lib/sync-engine";
@@ -212,12 +212,6 @@ export async function updateBloodPressureRecord(
 }
 
 // Pagination helpers
-
-export interface PaginatedResult<T> {
-  records: T[];
-  hasMore: boolean;
-  total: number;
-}
 
 export async function getWeightRecordsPaginated(
   page: number = 1,
